requirements.txt
torch # cpu版本也可
torchvision
mss
pillow
numpy<2
pynput
pyautogui